chore: adding our initial set up
