+++ .gitignore
@@ -191,3 +191,7 @@ build/
 # drawio temp files
 *.drawio.bkp
 *.drawio.dtmp
+
+#project specific
+**/node_modules
+**/dist
