--- apigateway/code/db.json
@@ -1,7 +1,0 @@
-{
-  "meta": {
-    "tile": "List of animals",
-    "date": "September 2024"
-  },
-  "animals" : []
-}
