--- apigateway/code/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "ex1",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "nodemon start.js"
-  },
-  "nodemonConfig": {
-    "ignore": [
-      "db.json"
-    ]
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.21.1",
-    "lowdb": "^7.0.1"
-  },
-  "devDependencies": {
-    "http-proxy-middleware": "^3.0.3",
-    "nodemon": "^3.1.7"
-  }
-}
-
-
-
-
