--- apigateway/code/routes/index.js
@@ -1,13 +1,0 @@
-import express from 'express';
-import { createProxyMiddleware } from 'http-proxy-middleware';
-const router = express.Router();
-
-// create a proxy for each microservice
-const microserviceProxy = createProxyMiddleware({
-  target: 'http://microservice:3011',
-  changeOrigin: true
-});
-
-router.use('/microservice', microserviceProxy);
-
-export default router;
