+++ apigateway/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "apigateway",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "dev": "concurrently -n \"TypeScript,Node\" -c \"blue,green\" \"npx tsc -w\" \"npx nodemon dist/start.js\""
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@types/express": "^4.17.21",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.2",
+    "express": "^4.17.1"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "concurrently": "^9.1.0",
+    "http-proxy-middleware": "^2.0.6",
+    "nodemon": "^3.0.1",
+    "typescript": "^5.6.3"
+  }
+}
