+++ apigateway/src/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "apigateway",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "type": "module",
+  "scripts": {
+    "dev": "./node_modules/nodemon/bin/nodemon.js start.js",
+    "dev win": "./node_modules/.bin/nodemon start.js"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@types/express": "^4.17.21",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.2",
+    "express": "^4.17.1"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "http-proxy-middleware": "^2.0.6",
+    "nodemon": "^3.0.1",
+    "typescript": "^5.6.3"
+  }
+}
