+++ apigateway/src/routes/index.ts
@@ -1,0 +1,10 @@
+import  express  from "express";
+import cors from 'cors';
+import {createProxyMiddleware } from 'http-proxy-middleware';
+const router = express.Router();
+
+
+
+
+
+export default router;
